pyproject.toml
[project]
name = "book_watch"
version = "0.1"
description = "Basic Python application for monitoring book availability on knihobot.cz."
authors = [
    {name = "Martin Barton", email = "ma.barton@seznam.cz"},
]
dependencies = [
    "requests>=2.31.0",
    "BeautifulSoup4>=4.12.2",
    "loguru>=0.7.2",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0 license"}

